endobelle-backend/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core>=1",
]

[project]
name = "endobella"
requires-python = ">=3.13"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
]

[tool.poetry]
name = "endobella"
version = "1.0.0"
description = ""
authors = [
  "Marcin Płotka <marcinplotka88@gmail.com>",
]

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.2.5"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
django-ckeditor-5 = "^0.2.18"
django-cors-headers = "^4.7.0"
django-environ = "^0.12.0"
django-database-url = "^1.0.3"
django-filter = "^25.1"
django-taggit = "6.1.0"
djoser = "^2.3.3"
pillow = "^11.3.0"
psycopg = "^3.2.9"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
gunicorn = "^23.0.0"
uvicorn = "^0.35.0"

pre-commit = "^4.3.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE="config.settings"
